API fixes from CI: packages, players accessor, shouldRemove, manual payload codecs; add temporary API probes

+++ src/main/java/com/tornadic/tornado/TornadoState.java
@@ -61,7 +61,7 @@ public class TornadoState {
 			ef = 0;
 		}
 		if (ef > peakEf) {
-			ef = peakEf;
+			ef = (int) peakEf;
 		}
 		return ef;
 	}
